+++ sales_management_mobile_app/WebAPI/Services/Director/DirectorServices.cs
@@ -100,7 +100,7 @@ namespace WebAPI.Services
             }
         }
 
-        public async Task<bool> createKpiValue(string dirId, int kpiValue)
+        public async Task<bool> createKpiValue(string dirId, decimal kpiValue)
         {
             //kpiValue cho Director
             var director = _context.Directors.Include(x => x.Managers).SingleOrDefault(x => x.Id == dirId);
@@ -110,13 +110,13 @@ namespace WebAPI.Services
             //cập nhật kpiYear và kpiValue cho Manager dười quyền Director và Salesman dười sự quản lý của Manager đó
             var managers = _context.Managers.Include(x => x.Users).Where(x => x.DirectorId == dirId).ToList();
             int countManagers = managers.Where(x => x.DirectorId == dirId).Count();
-            int kpiEachManager = kpiValue / countManagers;
+            decimal kpiEachManager = kpiValue / countManagers;
             for (int i = 0; i < countManagers; i++)
             {
                 managers[i].KpiValue = kpiEachManager;
                 var userss = _context.Users.Include(x => x.Stores).Where(x => x.ManagerId == managers[i].Id).ToList();
                 int countUser = userss.Where(x => x.ManagerId == managers[i].Id).Count();
-                int kpiEachUser = kpiEachManager / countUser;
+                decimal kpiEachUser = kpiEachManager / countUser;
                 for (int j = 0; j < countUser; j++)
                 {
                     userss[j].KpiValue = kpiEachUser;
